+++ client/packages/common/src/intl/locales/en/common.json
@@ -1196,8 +1196,7 @@
   "label.prescription-date": "Prescription Date",
   "label.previous-encounters": "Previous Encounters",
   "label.price": "Price",
-  "label.price-total": "Total Price",
-  "label.price-total-after-discount": "Total Price after Discount",
+  "label.cost-total-after-discount": "Total discounted cost",
   "label.pricing": "Pricing",
   "label.primary": "Primary",
   "label.primary-1": "Primary (1)",
