+++ client/packages/purchasing/src/purchase_order/api/operations.generated.ts
@@ -69,6 +69,8 @@ export type PurchaseOrderFragment = {
   supplierAgent?: string | null;
   supplierDiscountAmount: number;
   supplierDiscountPercentage?: number | null;
+  lineTotalAfterDiscount: number;
+  orderTotalAfterDiscount: number;
   targetMonths?: number | null;
   confirmedDatetime?: string | null;
   contractSignedDate?: string | null;
@@ -177,6 +179,8 @@ export type PurchaseOrderByIdQuery = {
         supplierAgent?: string | null;
         supplierDiscountAmount: number;
         supplierDiscountPercentage?: number | null;
+        lineTotalAfterDiscount: number;
+        orderTotalAfterDiscount: number;
         targetMonths?: number | null;
         confirmedDatetime?: string | null;
         contractSignedDate?: string | null;
@@ -519,6 +523,8 @@ export const PurchaseOrderFragmentDoc = gql`
     supplierAgent
     supplierDiscountAmount
     supplierDiscountPercentage
+    lineTotalAfterDiscount
+    orderTotalAfterDiscount
     targetMonths
     confirmedDatetime
     contractSignedDate
