+++ client/packages/purchasing/src/purchase_order/api/operations.graphql
@@ -92,6 +92,8 @@ fragment PurchaseOrder on PurchaseOrderNode {
   supplierAgent
   supplierDiscountAmount
   supplierDiscountPercentage
+  lineTotalAfterDiscount
+  orderTotalAfterDiscount
   targetMonths
   confirmedDatetime
   contractSignedDate
